added some comments and const

+++ src/crepe/api/CircleCollider.h
@@ -14,10 +14,10 @@ namespace crepe {
 class CircleCollider : public Collider {
 public:
 
-	CircleCollider(game_object_id_t game_object_id,vec2 offset, int radius);
+	CircleCollider(game_object_id_t game_object_id,vec2 offset, float radius);
 
 	//! Radius of the circle collider.
-	double radius;
+	float radius;
 };
 
 } // namespace crepe
